+++ catowseriOS/Domain/Sources/CottonUseCases/AppErrors.swift
@@ -10,14 +10,22 @@ import CottonTabs
 import CottonSearch
 import Foundation
 
+// MARK: - Errors
+
 /// Errors used on use case level
 public enum AppError: LocalizedError {
+    /// Use case reference is nil
     case zombieSelf
+    /// Search data service failure
     case searchDataServiceError(SearchServiceError)
+    /// Erased search data service failure
     case erasedSearchDataServiceError(Error)
+    /// Same command not finished yet
     case commandNotFinishedYet
+    /// Tabs data service failure
     case tabsServiceError(TabsListError)
     
+    /// Error description for localized error
     public var errorDescription: String? {
         switch self {
         case .zombieSelf:
