+++ catowseriOS/Domain/Sources/CottonUseCases/BaseUseCase.swift
@@ -6,6 +6,8 @@
 //  Copyright © 2024 Cotton (former Catowser). All rights reserved.
 //
 
+// MARK: - Protocol
+
 /// A use case type marker protocol to be able to filter classes and find only use cases if needed.
 ///
 /// In the future must be replaced on `CoreUseCase`.
